\name{addVariable}
\alias{addVariable}
\title{Function to create a formula which adds a variable to an input formula...}
\usage{addVariable(theFormula, addVar)}
\description{Function to create a formula which adds a variable to an input formula}
\value{The output is a formula object which is the input formula added to the variable}
\note{This should really be an internal function}
\keyword{formula}
\keyword{add}
\author{Chibisi Chima-Okereke \email{cchima-okereke@mango-solutions.com}}
\arguments{\item{\code{theFormula}}{this is the formula that the variable should be added to}
\item{\code{addVar}}{this is the variable that you should add}}
