\name{policyTable}
\title{This is a simulated dataset, it is a modified version of that given by Kaas, R., Goovaerts, M., 
Dhaene, J., Denuit, M. in Modern  Actuarial Risk Theory, and can be found at 
\url{http://www1.fee.uva.nl/ke/act/people/kaas/Cars.txt}}
\description{This data will be included the package}
\alias{autoPricing}
\docType{data}
\author{Chibisi Chima-Okereke \email{cchima-okereke@mango-solutions.com}}
\references{the data is an altered  version of the set used in KASS et al Modern 
Actuarial Risk Theory located at \url{http://www1.fee.uva.nl/ke/act/people/kaas/Cars.txt}}
\keyword{glm,}
\keyword{actuarial}
\keyword{pricing,}
\keyword{automating,}
\keyword{data}
\seealso{\code{\link{stepIC}}}
